Create Simple Mutex Example
This code introduces mutex locks. These ensure that a critical section of code is run 'in isolation'. For instance, if a thread modifies a variable that has global scope it should more than likely be encapsulated in a mutex lock while this modification is taking place. You may also want to lock something that reads in data to ensure the data isn't being modified while its being read.