Create PingPong
We can do slightly more with mutex locks other than just ensuring data or variables are good. We can use them to do coherent communications through the use of a global variable between threads. This program is one example of that sort of communication scheme. A 'ping' is sent from thread A to thread B, and a response 'pong' is sent back. We are of course simply locking a variable and ensuring its 'good' still, but the way we use that variable affords us more power and flexibility with the rest of our program. Notice we aren't actually sending or receiving messages but just writing to a variable and reading out of that same variable over and over.