Create False Sharing and its Simple(ish) fix
This is something you need to be aware of and is important. I am not entirely sure where I saw this originally anymore. I have had notes on this since I took the class you are now taking and the example I was given was very similar. We are not going to deal with this at the low level your book does. The short and simple version is that when your read data you have to be careful about the 'chunk' size of your data. If you overlap those chunks it leads to a performance loss and if you do it a lot with a lot of data the total performance loss can be quite staggering. You can fix this by simply padding your data so that when you read a chunk of data it never overlaps.